Fix bug introduced in b866ad85ce3b848a84c1912b7261ed4bd3f10eeb, Resolve MISRA 10.4 non-compliance

+++ embann.c
@@ -105,11 +105,11 @@ static int embann_initHiddenLayer(uint16_t numHiddenNeurons,
 
         for (uint16_t j = 0; j < numHiddenNeurons; j++)
         {    
-            wNeuron_t* pNeuron = (wNeuron_t*) malloc(sizeof(wNeuron_t) + (sizeof(neuronParams_t*) * (i == (0U ? numInputNeurons : numHiddenNeurons))));
+            wNeuron_t* pNeuron = (wNeuron_t*) malloc(sizeof(wNeuron_t) + (sizeof(neuronParams_t*) * (uint16_t)((i == 0U) ? numInputNeurons : numHiddenNeurons)));
             hiddenLayer->neuron[j] = pNeuron;
             hiddenLayer->neuron[j]->activation = 0.0F;
 
-            for (uint16_t k = 0; k < (i == (0U ? numInputNeurons : numHiddenNeurons)); k++)
+            for (uint16_t k = 0; k < (uint16_t)((i == 0U) ? numInputNeurons : numHiddenNeurons); k++)
             {
                 neuronParams_t* hiddenLayerParams = (neuronParams_t*) malloc(sizeof(neuronParams_t));
                 CHECK_MALLOC(hiddenLayerParams);
